+++ changelog.md
@@ -4,6 +4,8 @@
 
 - Now honoring the "active" flag on cycles imported from DPaintJS.
 - Better error reporting for opening files.
+- You can now load overlay pixels. They assume index 0 is transparent.
+- You can now collapse and expand the palette controls, as well as the new overlay controls. They start off collapsed by default.
 
 3: